refactor(models): split backend providers into familiar_runtime.models
Move the six provider implementations (Anthropic, OpenAI-compatible, Kimi,
GLM, Gemini, CLI) and their shared prompt-tool helpers out of the
1438-line src/familiar_agent/backend.py into dedicated modules under
src/familiar_runtime/models/. familiar_agent/backend.py shrinks to a
~200-line compatibility shell that re-exports the runtime adapters and
hosts the three configuration-driven factories (create_backend,
create_utility_backend, create_scene_backend).

This completes the PR 2 step of the agent-runtime reorganisation: model
provider code now lives behind the runtime package boundary, so the
runtime substrate no longer has any provider-specific code mixed with
neighbour-agent concerns. Legacy imports continue to work (TurnResult,
ToolCall, every backend class, and the private helpers _supports_adaptive_thinking,
_build_tools_system, _parse_tool_calls_from_text remain importable from
familiar_agent.backend).

Adds tests/test_runtime_models.py with provider-level smoke tests:
- shared helper behaviour (adaptive thinking detection, tool prompt
  injection, <tool_call> parsing)
- per-provider message factory serialisation (Anthropic / OpenAI native
  & prompt mode / Kimi / GLM / Gemini / CLI)
- backend.create_backend dispatching across all PLATFORM values
